Refactor schema logic in to type logic
